refactor: command 패키지 분리
