+++ src/main/java/command/Command.java
@@ -1,4 +1,4 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
