+++ src/main/java/command/End.java
@@ -1,27 +1,27 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
 
-public class Start implements Command {
+public class End implements Command {
 
-    private static final String START_COMMAND = "start";
+    private static final String END_COMMAND = "end";
 
-    private Start() {
+    private End() {
     }
 
-    public static Start from(String command) {
+    public static End from(String command) {
         validate(command);
-        return new Start();
+        return new End();
     }
 
     private static void validate(String command) {
-        if (!START_COMMAND.equals(command)) {
+        if (!END_COMMAND.equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
 
     @Override
     public void process(ChessGame chessGame) {
-        chessGame.start();
+        chessGame.end();
     }
 }
