+++ src/main/java/command/Load.java
@@ -1,27 +1,27 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
 
-public class End implements Command {
+public class Load implements Command {
 
-    private static final String END_COMMAND = "end";
+    private static final String LOAD_COMMAND = "load";
 
-    private End() {
+    private Load() {
     }
 
-    public static End from(String command) {
+    public static Load from(String command) {
         validate(command);
-        return new End();
+        return new Load();
     }
 
     private static void validate(String command) {
-        if (!END_COMMAND.equals(command)) {
+        if (!LOAD_COMMAND.equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
 
     @Override
     public void process(ChessGame chessGame) {
-        chessGame.end();
+        chessGame.load();
     }
 }
