+++ src/main/java/command/Save.java
@@ -1,4 +1,4 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
