+++ src/main/java/command/Start.java
@@ -1,26 +1,27 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
 
-public class Status implements Command {
+public class Start implements Command {
 
-    private static final String STATUS_COMMAND = "status";
+    private static final String START_COMMAND = "start";
 
-    private Status() {
+    private Start() {
     }
 
-    public static Status from(String command) {
+    public static Start from(String command) {
         validate(command);
-        return new Status();
+        return new Start();
     }
 
     private static void validate(String command) {
-        if (!STATUS_COMMAND.equals(command)) {
+        if (!START_COMMAND.equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
 
     @Override
     public void process(ChessGame chessGame) {
+        chessGame.start();
     }
 }
