+++ src/main/java/command/Status.java
@@ -1,27 +1,26 @@
-package domain.game.command;
+package command;
 
 import domain.game.ChessGame;
 
-public class Load implements Command {
+public class Status implements Command {
 
-    private static final String LOAD_COMMAND = "load";
+    private static final String STATUS_COMMAND = "status";
 
-    private Load() {
+    private Status() {
     }
 
-    public static Load from(String command) {
+    public static Status from(String command) {
         validate(command);
-        return new Load();
+        return new Status();
     }
 
     private static void validate(String command) {
-        if (!LOAD_COMMAND.equals(command)) {
+        if (!STATUS_COMMAND.equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
 
     @Override
     public void process(ChessGame chessGame) {
-        chessGame.load();
     }
 }
